refactor: share subscription source controller
