+++ packages/ui/src/product/converter/quick-mode/sources-section.tsx
@@ -1,4 +1,3 @@
-import * as React from "react";
 import * as Popover from "@radix-ui/react-popover";
 import { Plus, X, AlertCircle, Check, Loader2, Maximize2, HelpCircle, Menu, ChevronUp, ChevronDown } from "lucide-react";
 import { Button } from "@subboost/ui/components/ui/button";
@@ -8,249 +7,34 @@ import { Badge } from "@subboost/ui/components/ui/badge";
 import { Dialog, DialogContent, DialogHeader, DialogTitle } from "@subboost/ui/components/ui/dialog";
 import { Switch } from "@subboost/ui/components/ui/switch";
 import { cn } from "@subboost/ui/lib/utils";
-import { DEFAULT_NODE_NAME_TEMPLATE, formatNodeNameFromTemplate } from "@subboost/core/node-name-template";
-import type { ParsedNode } from "@subboost/core/types/node";
-import { normalizeSubscriptionImportErrorInfo } from "@subboost/core/subscription/import-error";
-import { getNodeSourceIds, useConfigStore, type SourceType, type SubscriptionSource } from "@subboost/ui/store/config-store";
-import { useUserStore } from "@subboost/ui/store/user-store";
-import { toast } from "@subboost/ui/components/ui/toaster";
+import { DEFAULT_NODE_NAME_TEMPLATE } from "@subboost/core/node-name-template";
+import type { SourceType } from "@subboost/ui/store/config-store";
 import { SubscriptionImportErrorBadge } from "@subboost/ui/product/converter/subscription-import-error";
 import { getSubscriptionUserInfoDisplay } from "@subboost/ui/product/subscription/subscription-userinfo-display";
-import { markSourceAsPendingImport } from "@subboost/ui/product/subscription/source-import-state";
-import { moveSubscriptionSource } from "@subboost/ui/product/subscription/source-order";
 import { buildSourceDisplayLabel } from "@subboost/ui/product/converter/source-display-label";
-import {
-  useProductInteractionAdapter,
-  type ProductInteractionResult,
-} from "@subboost/ui/product/interactions";
+import { useSubscriptionSourcesController } from "@subboost/ui/product/converter/use-subscription-sources-controller";
 import { sourceTypeInfo } from "./constants";
 
 export function SourcesSection() {
-  const [showAddMenu, setShowAddMenu] = React.useState(false);
-
-  const { nodes, parseErrors, sources, setSources, parseSingleSource } = useConfigStore();
-  const { user } = useUserStore();
-  const interactions = useProductInteractionAdapter();
-
-  const [expandedSourceId, setExpandedSourceId] = React.useState<string | null>(null);
-  const expandedSource = React.useMemo(
-    () => sources.find((s) => s.id === expandedSourceId) ?? null,
-    [expandedSourceId, sources]
-  );
-  const [expandedSourceSnapshot, setExpandedSourceSnapshot] = React.useState<{
-    id: string;
-    content: string;
-    tag: string;
-    nameTemplate: string;
-    useProxyProviders: boolean;
-    userinfoUrl: string;
-    userinfoUserAgent: string;
-  } | null>(null);
-
-  React.useEffect(() => {
-    if (!expandedSource) {
-      setExpandedSourceSnapshot(null);
-      return;
-    }
-    setExpandedSourceSnapshot((prev) => {
-      if (prev?.id === expandedSource.id) return prev;
-      return {
-        id: expandedSource.id,
-        content: expandedSource.content,
-        tag: (expandedSource.tag ?? "").trim(),
-        nameTemplate: (expandedSource.nameTemplate ?? "").trim(),
-        useProxyProviders: Boolean(expandedSource.useProxyProviders),
-        userinfoUrl: (expandedSource.userinfoUrl ?? "").trim(),
-        userinfoUserAgent: (expandedSource.userinfoUserAgent ?? "").trim(),
-      };
-    });
-  }, [expandedSource]);
-
-  const expandedSourcePreviewName = React.useMemo(() => {
-    if (!expandedSource) return "";
-    return formatNodeNameFromTemplate({
-      originName: "节点名称",
-      tag: expandedSource.tag,
-      template: expandedSource.nameTemplate,
-    });
-  }, [expandedSource]);
-
-  const closeExpandedSourceEditor = React.useCallback(() => {
-    if (expandedSource && expandedSourceSnapshot?.id === expandedSource.id && !expandedSource.parsing) {
-      const next = {
-        content: expandedSource.content,
-        tag: (expandedSource.tag ?? "").trim(),
-        nameTemplate: (expandedSource.nameTemplate ?? "").trim(),
-        useProxyProviders: Boolean(expandedSource.useProxyProviders),
-        userinfoUrl: (expandedSource.userinfoUrl ?? "").trim(),
-        userinfoUserAgent: (expandedSource.userinfoUserAgent ?? "").trim(),
-      };
-      const changed =
-        next.content !== expandedSourceSnapshot.content ||
-        next.tag !== expandedSourceSnapshot.tag ||
-        next.nameTemplate !== expandedSourceSnapshot.nameTemplate ||
-        next.useProxyProviders !== expandedSourceSnapshot.useProxyProviders ||
-        next.userinfoUrl !== expandedSourceSnapshot.userinfoUrl ||
-        next.userinfoUserAgent !== expandedSourceSnapshot.userinfoUserAgent;
-      if (changed) void parseSingleSource(expandedSource.id);
-    }
-    setExpandedSourceId(null);
-  }, [expandedSource, expandedSourceSnapshot, parseSingleSource]);
-
-  const maxSourcesPerType = React.useMemo(() => {
-    if (user?.isAdmin) return Number.POSITIVE_INFINITY;
-    if (!user) return 2;
-    const raw = user.quota?.maxImportSourcesPerType;
-    return typeof raw === "number" && Number.isFinite(raw) && raw >= 0 ? Math.floor(raw) : 5;
-  }, [user]);
-
-  const nodeCount = nodes.length;
-  const error = React.useMemo(
-    () => normalizeSubscriptionImportErrorInfo(parseErrors[0] ?? null)?.message ?? null,
-    [parseErrors]
-  );
-  const nodesBySourceId = React.useMemo(() => {
-    const grouped = new Map<string, ParsedNode[]>();
-    for (const node of nodes) {
-      for (const sourceId of getNodeSourceIds(node)) {
-        const current = grouped.get(sourceId);
-        if (current) {
-          current.push(node);
-        } else {
-          grouped.set(sourceId, [node]);
-        }
-      }
-    }
-    return grouped;
-  }, [nodes]);
-
-  const addSource = (type: SourceType) => {
-    const used = sources.filter((s) => s.type === type).length;
-    if (used >= maxSourcesPerType) {
-      toast({
-        title: user ? `每种导入方式最多 ${maxSourcesPerType} 个` : "未登录用户每种导入方式最多 2 个（登录后可提升）",
-        variant: "warning",
-      });
-      setShowAddMenu(false);
-      return;
-    }
-    const newSource: SubscriptionSource = {
-      id: Date.now().toString(),
-      type,
-      content: "",
-      nameTemplate: DEFAULT_NODE_NAME_TEMPLATE,
-    };
-    setSources([...sources, newSource]);
-    interactions.sourceAdded?.({
-      mode: "quick",
-      sourceType: type,
-      sourceCount: sources.length + 1,
-    });
-    setShowAddMenu(false);
-  };
-
-  const removeSource = (id: string) => {
-    if (sources.length <= 1) return;
-    setSources(sources.filter((s) => s.id !== id));
-  };
-
-  const updateSource = (id: string, content: string) => {
-    setSources(
-      sources.map((s) => {
-        if (s.id !== id) return s;
-        if (s.content === content) return s;
-        return markSourceAsPendingImport({ ...s, content });
-      })
-    );
-  };
-
-  const updateSourceMeta = (id: string, patch: Partial<SubscriptionSource>) => {
-    setSources(
-      sources.map((s) => {
-        if (s.id !== id) return s;
-
-        const next = { ...s, ...patch };
-        const changed = (Object.keys(patch) as Array<keyof SubscriptionSource>).some((key) => s[key] !== next[key]);
-        if (!changed) return s;
-
-        const needsReimport =
-          Object.prototype.hasOwnProperty.call(patch, "tag") ||
-          Object.prototype.hasOwnProperty.call(patch, "nameTemplate") ||
-          Object.prototype.hasOwnProperty.call(patch, "useProxyProviders") ||
-          Object.prototype.hasOwnProperty.call(patch, "userinfoUrl") ||
-          Object.prototype.hasOwnProperty.call(patch, "userinfoUserAgent");
-
-        return needsReimport ? markSourceAsPendingImport(next) : next;
-      })
-    );
-  };
-
-  const moveSource = React.useCallback(
-    (sourceId: string, direction: "up" | "down") => {
-      const nextSources = moveSubscriptionSource(sources, sourceId, direction);
-      if (nextSources === sources) return;
-      setSources(nextSources);
-    },
-    [setSources, sources]
-  );
-
-  const updateSourceType = (id: string, type: SourceType) => {
-    const current = sources.find((s) => s.id === id);
-    if (!current) return;
-    if (current.type === type) return;
-
-    const used = sources.filter((s) => s.type === type).length;
-    if (used >= maxSourcesPerType) {
-      toast({
-        title: user ? `每种导入方式最多 ${maxSourcesPerType} 个` : "未登录用户每种导入方式最多 2 个（登录后可提升）",
-        variant: "warning",
-      });
-      return;
-    }
-
-    setSources(
-      sources.map((s) => {
-        if (s.id !== id) return s;
-
-        return markSourceAsPendingImport({
-          ...s,
-          type,
-          content: "",
-          useProxyProviders: type === "url" ? Boolean(s.useProxyProviders) : undefined,
-          userinfoUrl: type === "url" ? s.userinfoUrl : undefined,
-          userinfoUserAgent: type === "url" ? s.userinfoUserAgent : undefined,
-          lastParsedContent: undefined,
-          lastParsedTag: undefined,
-          lastParsedNameTemplate: undefined,
-        });
-      })
-    );
-  };
-
-  const handleImportSource = React.useCallback(
-    async (sourceId: string) => {
-      const source = sources.find((item) => item.id === sourceId);
-      if (!source || !source.content.trim() || source.parsing) return;
-
-      await parseSingleSource(sourceId);
-
-      const latestSource = useConfigStore.getState().sources.find((item) => item.id === sourceId) ?? source;
-      const result: ProductInteractionResult = latestSource.parsed
-        ? "success"
-        : latestSource.error || latestSource.errorInfo
-          ? "runtimeError"
-          : "validationError";
-      interactions.sourceImported?.({
-        mode: "quick",
-        sourceType: latestSource.type,
-        result,
-        sourceCount: useConfigStore.getState().sources.filter((item) => item.content.trim()).length,
-        nodeCount: latestSource.nodeCount ?? 0,
-        usesProxyProvider: Boolean(latestSource.useProxyProviders),
-      });
-    },
-    [interactions, parseSingleSource, sources],
-  );
+  const {
+    addSource,
+    closeExpandedSourceEditor,
+    error,
+    expandedSource,
+    expandedSourcePreviewName,
+    handleImportSource,
+    moveSource,
+    nodeCount,
+    nodesBySourceId,
+    removeSource,
+    setExpandedSourceId,
+    setShowAddMenu,
+    showAddMenu,
+    sources,
+    updateSource,
+    updateSourceMeta,
+    updateSourceType,
+  } = useSubscriptionSourcesController({ mode: "quick" });
 
   return (
     <>
